Remove command that are not implemented
